1. Changed prefix type in Hash and Merge traits to u64. 2. MerkleMountainRange::new() now calculate num_leaves. Also changed num_leaves from Option<usize> to usize. 3. Fixed MerkleMountainRange::find(). 4. Created MerkleMountainRange::get() and MerkleMountainRange::get_leaf() to get elements from the tree.

+++ src/hash.rs
@@ -1,8 +1,8 @@
 pub trait Hash<H: Merge> {
-    fn hash(&self, prefix: usize) -> H;
+    fn hash(&self, prefix: u64) -> H;
 }
 
 pub trait Merge {
-    fn empty(prefix: usize) -> Self;
-    fn merge(&self, other: &Self, prefix: usize) -> Self;
+    fn empty(prefix: u64) -> Self;
+    fn merge(&self, other: &Self, prefix: u64) -> Self;
 }
