+++ src/mmr/proof.rs
@@ -36,7 +36,7 @@ impl<H: Merge + Clone> Proof<H> {
                 if pos.index >= self.mmr_size {
                     return Err(Error::ProveInvalidLeaves);
                 }
-                let hash = value.hash(pos.num_leaves());
+                let hash = value.hash(pos.num_leaves() as u64);
                 Ok((pos, hash))
             })
             .collect();
@@ -145,9 +145,9 @@ impl<H: Merge + Clone> Proof<H> {
 
             let parent_pos = position.parent();
             let parent_hash = if position.right_node {
-                sibling_hash.merge(&hash, parent_pos.num_leaves())
+                sibling_hash.merge(&hash, parent_pos.num_leaves() as u64)
             } else {
-                hash.merge(&sibling_hash, parent_pos.num_leaves())
+                hash.merge(&sibling_hash, parent_pos.num_leaves() as u64)
             };
 
             // Add the parent to the queue.
